// src/features/projects/types.ts
// export type Project = {
//   id: string;
//   slug: string;
//   name: string;
//   description: string;
//   color: string; // "bg-pink-500" etc
//   status: "active" | "archived";
//   tasksTotal: number;
//   tasksCompleted: number;
//   dueDate: string;
// };

export type Project = {
  id: string;
  slug: string;
  name: string;
  description: string;
  color: string; // "bg-pink-500"
  status: "active" | "archived";
  tasksTotal: number;
  tasksCompleted: number;
  dueDate: string;
};

export type Task = {
  id: string;
  title: string;
  status: "todo" | "in_progress" | "done";
};